requirements.txt
selenium
PyMuPDF
gtts
google-api-python-client
django
django-cors-headers
djangorestframework
python-dotenv
pymupdf
cloudinary
cryptography
PyJWT
